feat: use showcase list for homepage projects
The landing page still used the old card grid. Match /projects and drop the unused card components.
